refactor uploadPhoto to ProfilePhoto
