+++ server/app.js
@@ -12,7 +12,7 @@ const logger = require("morgan");
 
 const authRouter = require("./routes/auth");
 const userRouter = require("./routes/user");
-const uploadRouter = require("./routes/upload");
+const photoRouter = require("./routes/photo");
 const profileRouter = require("./routes/profile");
 const requestRouter = require("./routes/request");
 
@@ -48,7 +48,7 @@ app.use((req, res, next) => {
 app.use("/auth", authRouter);
 app.use("/users", userRouter);
 app.use("/profiles", profileRouter);
-app.use("/api", uploadRouter);
+app.use("/api", photoRouter);
 app.use("/requests", requestRouter);
 
 if (process.env.NODE_ENV === "production") {
